# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/acls": {
            "get": {
                "description": "Get all ACLs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Acl"
                ],
                "summary": "Get all ACLs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "pagesize",
                        "name": "pagesize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "pagenum",
                        "name": "pagenum",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Acl"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/controllers.errResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.errResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a acl",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Acl"
                ],
                "summary": "Create a acl",
                "parameters": [
                    {
                        "description": "填写ACL信息",
                        "name": "AclInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Acl"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "comment": {
                                            "type": "string"
                                        },
                                        "disabled": {
                                            "type": "boolean"
                                        },
                                        "filter": {
                                            "type": "string"
                                        },
                                        "name": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Acl"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/controllers.errResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.errResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/acls/{id}": {
            "get": {
                "description": "Get acl by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Acl"
                ],
                "summary": "Get acl by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "acl id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Acl"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controllers.errResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.errResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Create a acl",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Acl"
                ],
                "summary": "Create a acl",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "填写Acl信息",
                        "name": "aclInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controllers.aclCreateForm"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "comment": {
                                            "type": "string"
                                        },
                                        "disabled": {
                                            "type": "boolean"
                                        },
                                        "name": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Acl"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.errResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Get acl by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Acl"
                ],
                "summary": "Get acl by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{error:{code:400,message:\"bad request\"}}",
                        "schema": {
                            "$ref": "#/definitions/controllers.errResponse"
                        }
                    },
                    "404": {
                        "description": "{error:{code:404,message:\"record not found\"}}",
                        "schema": {
                            "$ref": "#/definitions/controllers.errResponse"
                        }
                    },
                    "500": {
                        "description": "{error:{code:404,message:\"server error\"}}",
                        "schema": {
                            "$ref": "#/definitions/controllers.errResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/ipdb": {
            "get": {
                "description": "Get all IPDB",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IPDB"
                ],
                "summary": "Get all IPDB",
                "parameters": [
                    {
                        "type": "string",
                        "description": "pagesize",
                        "name": "pagesize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "pagenum",
                        "name": "pagenum",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.IPDB"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/controllers.errResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.errResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a ipdb",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IPDB"
                ],
                "summary": "Create a ipdb",
                "parameters": [
                    {
                        "description": "填写ip信息",
                        "name": "IPDBInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.IPDB"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.IPDB"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/controllers.errResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.errResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controllers.aclCreateForm": {
            "type": "object",
            "required": [
                "name",
                "path"
            ],
            "properties": {
                "comment": {
                    "type": "string"
                },
                "disabled": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                }
            }
        },
        "controllers.errResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.Acl": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "comment": {
                    "type": "string"
                },
                "disabled": {
                    "type": "boolean"
                },
                "filter": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.IPDB": {
            "type": "object",
            "properties": {
                "cidr": {
                    "type": "string"
                },
                "comment": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isp": {
                    "type": "string"
                },
                "province": {
                    "type": "string"
                }
            }
        },
...
    }
}